+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # VRFSearchTool.py #
 ---
 
+## VRFSearchTool 0.06 (2013-07-25) ##
+* Alpha release of the tool.  Still requires some error checking, code cleanup
+  and commenting but otherwise is a working application.
+
 ## VRFSearchTool 0.01 (2013-07-12) ##
-* Initial commit.
+* Initial commit of support files